move cfg conversion to parser.py
