+++ tdmpc2/train.py
@@ -18,42 +18,9 @@ from tdmpc2 import TDMPC2
 from trainer.offline_trainer import OfflineTrainer
 from trainer.online_trainer import OnlineTrainer
 from common.logger import Logger
-import dataclasses
-from typing import Any
-from omegaconf import OmegaConf
 torch.backends.cudnn.benchmark = True
-
 torch.set_float32_matmul_precision('high')
 
-def cfg_to_dataclass(cfg, frozen=False):
-	# Converts an OmegaConf config to a dataclass, which will not cause graph breaks
-	cfg_dict = OmegaConf.to_container(cfg)
-	fields = []
-	for key, value in cfg_dict.items():
-		fields.append((key, Any, dataclasses.field(default_factory=lambda value_=value: value_)))
-
-	# Create the dataclass
-	dataclass_name = "Config"
-	dataclass = dataclasses.make_dataclass(dataclass_name, fields, frozen=frozen)
-	def get(self, val, default=None):
-		return getattr(self, val, default)
-	dataclass.get = get
-	return dataclass()
-
-def cfg_to_dataclass(cfg, frozen=False):
-	# Converts an OmegaConf config to a dataclass, which will not cause graph breaks
-	cfg_dict = OmegaConf.to_container(cfg)
-	fields = []
-	for key, value in cfg_dict.items():
-		fields.append((key, Any, dataclasses.field(default_factory=lambda value_=value: value_)))
-
-	# Create the dataclass
-	dataclass_name = "Config"
-	dataclass = dataclasses.make_dataclass(dataclass_name, fields, frozen=frozen)
-	def get(self, val, default=None):
-		return getattr(self, val, default)
-	dataclass.get = get
-	return dataclass()
 
 @hydra.main(config_name='config', config_path='.')
 def train(cfg: dict):
@@ -82,9 +49,6 @@ def train(cfg: dict):
 	print(colored('Work dir:', 'yellow', attrs=['bold']), cfg.work_dir)
 
 	trainer_cls = OfflineTrainer if cfg.multitask else OnlineTrainer
-
-	cfg = cfg_to_dataclass(cfg)
-
 	trainer = trainer_cls(
 		cfg=cfg,
 		env=make_env(cfg),
